🚚 Move: `ErrorResponse` struct to `model::response` module [#114]
